services/accounts/requirements.txt
alembic==1.0.5
appnope==0.1.0
backcall==0.1.0
Click==7.0
decorator==4.3.0
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
grpcio==1.17.1
grpcio-tools==1.17.1
ipython==7.2.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.3.1
python-dateutil==2.7.5
python-editor==1.0.3
six==1.12.0
SQLAlchemy==1.2.15
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
